Fixed Fisheye but Ugly walls

+++ srcs/main.c
@@ -6,7 +6,7 @@
 /*   By: mratsima <mratsima@student.42antananari    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: Invalid date        by                   #+#    #+#             */
-/*   Updated: 2025/11/12 12:49:17 by mratsima         ###   ########.fr       */
+/*   Updated: 2025/11/13 10:04:10 by mratsima         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -252,7 +252,7 @@ int		draw_line_2(t_display *display, float beta)
 {
 	t_line *tmp;
 	t_point tmp_bloc;
-	int	distance;
+	float	distance;
 
 	tmp = display->head;
 	while (tmp)
